Refactor editor to block
